requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
aiohttp==3.8.5
async-socks==0.2.4
redis==4.6.0
pydantic==2.3.0
pydantic-settings==2.0.3
apscheduler==3.10.4
python-dotenv==1.0.0
asyncpg==0.28.0
python-multipart==0.0.6
httpx==0.24.1
jinja2==3.1.2
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
